REFACTOR: Groceries and add modals
