+++ src/client/Pages/Groceries/index.tsx
@@ -1,11 +1,13 @@
+import { useModal } from 'client/hooks/useModal'
 import { useGetGroceriesQuery } from 'client/Redux/api/groceries'
 import { Grocery } from 'model/types'
 import React from 'react'
 import { useNavigate } from 'react-router-dom'
-import AddForm from './AddForm'
+import AddGroceryModal from './AddGroceryModal'
 
 const Groceries = () => {
   const navigate = useNavigate()
+  const addGroceryModal = useModal()
 
   const { data: groceries, isLoading } = useGetGroceriesQuery(undefined, {
     refetchOnMountOrArgChange: true,
@@ -14,10 +16,10 @@ const Groceries = () => {
   return (
     <div>
       <h1>Groceries</h1>
-      <AddForm />
-      <h2>Listing</h2>
+      <button onClick={addGroceryModal.show}>Add</button>
+      <AddGroceryModal modal={addGroceryModal} />
       {isLoading && <p>Loading...</p>}
-      {groceries && groceries.length > 0 && (
+      {groceries && groceries.length > 0 ? (
         <ul>
           {groceries.map((grocery: Grocery) => (
             <li key={grocery.id}>
@@ -34,6 +36,8 @@ const Groceries = () => {
             </li>
           ))}
         </ul>
+      ) : (
+        <p>No groceries found.</p>
       )}
     </div>
   )
